% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Paris20eAddresses}
\alias{Paris20eAddresses}
\title{Voter addresses in the 20th arrondissement of Paris}
\format{
An object of class \code{sf} (inherits from \code{grouped_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5820 rows and 3 columns.
}
\source{
INSEE: \url{https://www.data.gouv.fr/datasets/bureaux-de-vote-et-adresses-de-leurs-electeurs}
}
\usage{
Paris20eAddresses
}
\description{
All voter addresses located in the 20th arrondissement of Paris, according to
the Répertoire électoral unique (REU). We first extracted polling stations
from \code{table-bv-reu.parquet}, and then extracted addresses located within
them from \code{table-adresses-reu.parquet}. Addresses were converted into
spatial points of class \code{sf}. The weight variable is \code{nb_adresses},
although that variable does not count voters per se: it measures the number
of voter addresses at that particular location, which is probably the best
publicly available proxy at that level of precision.
}
\keyword{datasets}
