% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ParisIris}
\alias{ParisIris}
\title{Paris IRIS shapefile, 2013}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 992 rows and 7 columns.
}
\source{
IGN: \url{https://cartes.gouv.fr/rechercher-une-donnee/dataset/IGNF_CONTOURS-IRIS}
}
\usage{
ParisIris
}
\description{
A shapefile of the IRIS (French census tract) of Paris in 2013.
}
\keyword{datasets}
