% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfAggregate.R
\name{sfAggregate}
\alias{sfAggregate}
\title{Aggregate Spatial Polygons}
\usage{
sfAggregate(sp, id)
}
\arguments{
\item{sp}{an \code{\link[sp]{SpatialPolygonsDataFrame}} object}

\item{id}{a grouping vector, as required by \code{\link[sf]{aggregate}}}
}
\value{
a \code{\link[sp]{SpatialPolygonsDataFrame}} object with unique and
valid feature IDs
}
\description{
A replacement for \code{rgeos::unaryUnion}.
}
\note{
Source: \url{https://github.com/r-spatial/sf/issues/2563}
}
\author{
Roger Bivand (thanks)
}
