% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfUnionSpatialPolygons.R
\name{sfUnionSpatialPolygons}
\alias{sfUnionSpatialPolygons}
\title{Aggregate Polygons in a SpatialPolygons object (using sf)}
\usage{
sfUnionSpatialPolygons(sp, id)
}
\arguments{
\item{sp}{a `SpatialPolygonsDataFrame` object}

\item{id}{a string identifying the polygons}
}
\description{
An attempt at merging polygons without \code{rgeos}, using \code{sf}, and
without losing too much speed. Very experimental, and very slow despite my
best efforts: see \url{https://github.com/r-spatial/sf/issues/2563} for
details on what I tried.
}
\author{
François Briatte
}
