% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfReapportion.R
\name{sfReapportion}
\alias{sfReapportion}
\title{Reapportion data from one geography to another}
\usage{
sfReapportion(
  old_geom,
  new_geom,
  data,
  old_ID,
  new_ID,
  data_ID,
  variables = names(data)[-which(names(data) \%in\% data_ID)],
  mode = "count",
  weights = NULL,
  weight_matrix = NULL,
  weight_matrix_var = NULL
)
}
\arguments{
\item{old_geom}{a `SpatialPolygonsDataFrame` or `sf` object
representing the initial geometry.}

\item{new_geom}{a `SpatialPolygonsDataFrame` or `sf` object
representing the geometry you want to reapportion data to.}

\item{data}{a `data.frame` containing the data to reapportion, and an ID
allowing to match it to the `old_geom`.}

\item{old_ID}{a string, the name of the ID variable in the `old_geom`.}

\item{new_ID}{a string, the name of the ID variable in the `new_geom`.}

\item{data_ID}{a string, the name of the ID variable in the `data`.}

\item{variables}{a character vector, representing the names of the variables
in the `data` set to reapportion. By default, all data variables except for
the ID.}

\item{mode}{either \code{"count"} or \code{"proportion"}. \code{"count"} is
for absolute values, \code{"proportion"} is for, well, proportions (expressed
between 0 and 1). If \code{"proportion"}, a `weights` variable needs to be
provided. \strong{Note: \code{"proportion"} has only been lightly tested.}}

\item{weights}{\strong{(optional, lightly tested only)} In case the variables
are proportions, the name of the variable containing weights (i.e. the total
number of observations per unit in the `old_geom`).}

\item{weight_matrix}{\strong{(optional, lightly tested only)} a
`SpatialPointsDataFrame` or `sf` object indicating the spatial coordinates of
the observations (inhabitants, voters, etc.).}

\item{weight_matrix_var}{\strong{(optional, lightly tested only)} the name of
the (numeric) variable containing the weights in \code{weight_matrix}.}
}
\value{
a `data.frame` containing \code{new_ID} and the reapportioned
\code{variables} from \code{data}
}
\description{
This function allows to reapportion data from one geography to another, for
example in the context of working with different administrative units.
}
\note{
Inspiration from \url{https://stackoverflow.com/a/17703903} and
\url{https://rpubs.com/PaulWilliamson/6577}.
Original \code{sp} version available at \url{https://github.com/joelgombin/spReapportion}.
All mistakes are mine, obviously. More details appear in the README file.
}
\author{
Joël Gombin (initial \code{sp} version),
François Briatte (\code{sf} port)
}
